+++ test/libsolidity/semanticTests/strings/concat/string_concat_empty_argument_list.sol
@@ -1,0 +1,10 @@
+contract C {
+    function f() public returns (string memory) {
+        return string.concat();
+    }
+}
+// ====
+// compileToEwasm: also
+// compileViaYul: also
+// ----
+// f() -> 0x20, 0